feat: RestService, update: GET entrypoint with sorted list
