+++ src/main/java/com/looqbox/challenge/model/dto/PokemonEntryDto.java
@@ -1,5 +1,6 @@
 package com.looqbox.challenge.model.dto;
 
+import com.looqbox.challenge.model.Pokemon;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,6 +9,6 @@ import lombok.Data;
 @AllArgsConstructor
 public final class PokemonEntryDto {
 
-  private final List<String> result;
+  private final List<Pokemon> result;
 
 }
